+++ test/configparser.cpp
@@ -95,7 +95,10 @@ TEST_CASE("\"unbind-key -a\" removes all key bindings", "[ConfigParser]")
 		cfgparser.parse("data/config-unbind-all");
 
 		for (int i = OP_QUIT; i < OP_NB_MAX; ++i) {
-			REQUIRE(keys.getkey(static_cast<Operation>(i), "all") == "<none>");
+			REQUIRE(keys.get_keys(static_cast<Operation>(i),
+					"feedlist") == std::vector<std::string>());
+			REQUIRE(keys.get_keys(static_cast<Operation>(i),
+					"podboat") == std::vector<std::string>());
 		}
 	}
 
@@ -108,11 +111,12 @@ TEST_CASE("\"unbind-key -a\" removes all key bindings", "[ConfigParser]")
 		KeyMap default_keys(KM_NEWSBOAT);
 		for (int i = OP_QUIT; i < OP_NB_MAX; ++i) {
 			const auto op = static_cast<Operation>(i);
-			REQUIRE(keys.getkey(op, "help") == default_keys.getkey(op, "help"));
+			REQUIRE(keys.get_keys(op, "help") == default_keys.get_keys(op, "help"));
 		}
 
 		for (int i = OP_QUIT; i < OP_NB_MAX; ++i) {
-			REQUIRE(keys.getkey(static_cast<Operation>(i), "article") == "<none>");
+			REQUIRE(keys.get_keys(static_cast<Operation>(i),
+					"article") == std::vector<std::string>());
 		}
 	}
 }